Title: 두 수의 합, Time: 231ms, Memory: 80MB, Status: Memory Limit Exceed - CodeTree

+++ 240902/두 수의 합/sum-of-two-num.cpp
@@ -3,7 +3,7 @@
 
 using namespace std;
 #define ll  long long
-map<int, int> m;
+map<int, short> m;
 
 ll factorial(int n) {
     if (n == 1)return 1;
